+++ WordFramework/Word+CoreDataProperties.swift
@@ -17,15 +17,15 @@ extension Word {
     }
 
     @NSManaged public var my_context: String
-    @NSManaged public var my_frequency: Int16
-    @NSManaged public var my_index: Int32
+    @NSManaged public var my_frequency: Int64
+    @NSManaged public var my_index: Int64
     @NSManaged public var my_level: Double
     @NSManaged public var my_meanings: [String]
     @NSManaged public var my_pinyin: String?
     @NSManaged public var my_simplified: String?
-    @NSManaged public var my_spokenFrequency: Int16
+    @NSManaged public var my_spokenFrequency: Int64
     @NSManaged public var my_traditional: String
-    @NSManaged public var my_writtenFrequency: Int16
+    @NSManaged public var my_writtenFrequency: Int64
     @NSManaged public var my_zhuyin: String
 
 }
